gameplay: state bytes and the second actor slot are members
Placeholders into Gp_StateC08 and Gp_ActorSlots[1] become member accesses;
field_9 is read signed everywhere, so it is s8, and two actors' undersized
Gp_ActorSlots[1] redeclarations go.

+++ src/actors/actor_444000/actor_444000.c
@@ -441,7 +441,7 @@ void func_actor_444000_80132358(Task* task)
     if (gGameSession->field_65 != 0) {
         return;
     }
-    if ((s8)Gp_StateC08.field_9 != 0) {
+    if (Gp_StateC08.field_9 != 0) {
         return;
     }
     if (Gp_StateF0.field_4 != 0) {
